Various bugfixes for MSVC build

+++ tests/geometry/2d/testSegmentComputerFunctor.cpp
@@ -86,7 +86,7 @@ bool testTangentFromDSS(
   TangentAngleFromDSSFunctor<DSSComputer> f; 
   //call
   double v1 = f(*begin,dss); 
-  double v2 = std::atan2(dss.getA(),dss.getB());
+  double v2 = std::atan2((double)dss.getA(),(double)dss.getB());
   trace.info() << "Tangent orientation : " << v1 << " == " << v2 << endl;
 
   return (v1 == v2);
